+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 -----------
 
 ### Internals
-* None.
+* Update tests to use global logger. ([PR #6917](https://github.com/realm/realm-core/pull/6917))
 
 ----------------------------------------------